Bluetooth: controller: Support BT 5.0 feature set bit fields
Added support in the Controller for increased feature set
bit fields in Bluetooth 5.0. Cached feature set in the
connection context is increased to a uint32_t value to
accommodate 17 feature bits.


Former-commit-id: 11a32bf46d0c2f2908ab082c501deb072b1429b6

+++ subsys/bluetooth/controller/ll_sw/ctrl_internal.h
@@ -72,7 +72,7 @@ DECL|latency_prepare|member|uint16_t latency_prepare;
 DECL|latency|member|uint16_t latency;
 DECL|latency|member|uint16_t latency;
 DECL|llcp_ack|member|uint8_t llcp_ack;
-DECL|llcp_features|member|uint8_t llcp_features;
+DECL|llcp_features|member|uint32_t llcp_features;
 DECL|llcp_length|member|} llcp_length;
 DECL|llcp_req|member|uint8_t llcp_req;
 DECL|llcp_terminate|member|} llcp_terminate;
